server: Share TOTP failure classification and extract MFA helpers
Put totp_failure next to the marker constants in lldap_mfa so web, LDAP,
and the login form cannot drift. Move access-control MFA policy helpers
and server login free functions into sibling modules to shrink the
diff in upstream-owned files.
