+++ crates/mfa/src/lib.rs
@@ -27,7 +27,8 @@ pub use totp::{format_code, split_totp_suffix, totp_code, totp_verify};
 pub use types::{SEALED_BLOB_LEN, SEALED_PREFIX};
 pub use types::{
     TOTP_CODE_ALREADY_USED, TOTP_CURRENT_CODE_REQUIRED, TOTP_DIGITS, TOTP_ENROLLMENT_EXPIRED,
-    TOTP_ENROLLMENT_TTL_SECS, TOTP_SEED_LEN, TOTP_SEPARATOR, TOTP_TOO_MANY_ATTEMPTS,
+    TOTP_ENROLLMENT_TTL_SECS, TOTP_SEED_LEN, TOTP_SEPARATOR, TOTP_TOO_MANY_ATTEMPTS, TotpFailure,
+    totp_failure,
 };
 #[cfg(feature = "seal")]
 pub use used_codes::UsedCodes;
